backend/requirements.txt
django>=5.0,<6.0
djangorestframework>=3.15.0
djangorestframework-simplejwt>=5.3.1
django-filter>=24.2
drf-spectacular>=0.27.2
psycopg2-binary>=2.9.9
python-decouple>=3.8
django-cors-headers>=4.4.0
django-storages>=1.14.4
boto3>=1.35.0
